requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
Pillow
joblib
catboost